Split some code in two modules taskList and todoTask
