+++ src/modules/todoTask.js
@@ -1,0 +1,7 @@
+export default class TodoTask {
+  constructor(desp, arr, done = false) {
+    this.desp = desp;
+    this.index = arr.length;
+    this.done = done;
+  }
+}
